Demo中去除QY
