+++ src/Senparc.Weixin.MP.Sample/Senparc.Weixin.MP.Sample/Senparc.Weixin.MP.Sample.csproj
@@ -172,7 +172,6 @@
     <Compile Include="Controllers\OAuth2Controller.cs" />
     <Compile Include="Controllers\Open\OpenController.cs" />
     <Compile Include="Controllers\Open\OpenOAuthController.cs" />
-    <Compile Include="Controllers\QY\QYController.cs" />
     <Compile Include="Controllers\SimulateToolController.cs" />
     <Compile Include="Controllers\TenPayV3Controller.cs">
       <SubType>Code</SubType>
